Base reorganization complete. Need to update references and verify applications are operational.
