--- src/qdlutils/applications/controllers/nidaq_wm_ple.yaml
@@ -1,39 +1,0 @@
-QT3PLE:
-  ApplicationController:
-    import_path : qdlutils.datagenerators.plescanner
-    class_name : PleScanner
-    configure :
-      controller: VoltageController
-      readers :
-        daq_readers:
-          daq_reader : DAQReader
-        wm_readers:
-          wm_reader : WavemeterDllController
-
-  DAQReader:
-    import_path : qdlutils.applications.controllers.lockin_controller
-    class_name  : Lockin
-    configure :
-      daq_name : Silicon_DAQ # NI DAQ Device Name
-      signal_channels : ai21  # NI DAQ analog input channels to use for reading position
-      sample_number: 20  # Specifies the sampling rate in samples per channel per second
-      rate: 20.0  #Specifies the number of samples to acquire or generate for each channel in the task
-      timeout: 20  # Specifies the amount of time in seconds to wait for samples to become available
-
-  VoltageController:
-    import_path : qdlutils.nidaq.customcontrollers
-    class_name  : VControl
-    configure :
-      daq_name : Silicon_DAQ # NI DAQ Device Name
-      write_channels : ao3 # NI DAQ analog output channels to use for writing voltage
-      read_channels : ai0  # NI DAQ analog input channels to use for reading voltage
-      min_position: -10  # conversion factor from volts to microns, can also supply a list [8,8,8] or [6,4.2,5]
-      max_position: 10  # the voltage value that defines the position 0,0,0, can also supply a list [0,0,0] or [5,5,5]
-      scale_nm_per_volt: 1  # microns
-      num_measurements_per_batch: 10
-
-  WavemeterDllController:
-    import_path : qdlutils.applications.controllers.wavemeter_controller
-    class_name  : WavemeterDllController
-    configure :
-      dll_path : C:/Users/Fulab/Downloads/00392-2-06-A_CustomerCD621/00392-2-06-A_Customer CD 621/Programming Interface/x64/CLDevIFace.dll
