+++ src/qdlutils/applications/qt3scan/config_files/andor_spectrometer.yaml
