+++ src/qdlutils/applications/qt3scan/config_files/princeton_spectrometer.yaml
