+++ src/qdlutils/applications/qt3scan/config_files/random_spectrometer.yaml
