+++ src/qdlutils/applications/qt3scope/config_files/random_data_generator.yaml
@@ -1,0 +1,9 @@
+QT3Scope:
+  DAQController:
+    import_path : qdlutils.applications.controllers.random_data_generator
+    class_name  : QT3ScopeRandomDataController
+    configure : 
+      simulate_single_light_source : False
+      num_data_samples_per_batch : 10
+      default_offset: 100
+      signal_noise_amp: 0.5
