--- src/qdlutils/datagenerators/__init__.py
@@ -1,5 +1,0 @@
-from .daqsamplers import RandomRateCounter
-from .daqsamplers import NiDaqDigitalInputRateCounter
-from .piezoscanner import CounterAndScanner
-from .piezoscanner import CounterAndScanner as NiDaqPiezoScanner
-from .piezoscanner import CounterAndScanner as RandomPiezoScanner
