+++ src/qdlutils/hardware/nidaq/config.py
