+++ src/qdlutils/hardware/nidaq/nidaqedgecounter.py
