+++ src/qdlutils/hardware/spectrometers/andor.py
