+++ src/qdlutils/hardware/spectrometers/spectrometer.py
